+++ src/components/shared/RegisterInputWrapper/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import INPUT_TYPE from '../../../constants/inputType';
+import { INPUT_TYPE } from '../../../constants/constants';
 import * as Style from './style';
 
 const RegisterInputWrapper = (props) => {
@@ -9,12 +9,9 @@ const RegisterInputWrapper = (props) => {
   const isOwnerName = type === INPUT_TYPE.OWNER_NAME.type;
   const wordCount = <Style.WordCount>{`${currentNameLength}/30`}</Style.WordCount>;
 
-  const isCardPassword = type === INPUT_TYPE.PASSWORD.type;
-
   return (
     <Style.Root>
-      {/* TODO: cardPassword일 때 inputHeader width 조정하기 */}
-      <Style.InputHeader width={isCardPassword ? '100%' : width}>
+      <Style.InputHeader width={width}>
         <Style.Label htmlFor={type}>{label}</Style.Label>
         {isOwnerName && wordCount}
       </Style.InputHeader>
